removed moved util functions to application
